pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bellwether"
version = "0.1.0"
description = "The cost-and-failure-mode benchmark for LLM agents."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
  { name = "Stephen Hedrick", email = "Stephen@wavebound.io" },
]
keywords = [
  "llm",
  "benchmark",
  "evaluation",
  "agents",
  "anthropic",
  "openai",
  "gemini",
  "tcot",
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  "anthropic>=0.40",
  "openai>=1.50",
  "google-genai>=0.3",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=4.1",
  "ruff>=0.5",
  "mypy>=1.10",
]

[project.urls]
Homepage = "https://github.com/cartesianxr7/bellwether"
Repository = "https://github.com/cartesianxr7/bellwether"
Documentation = "https://cartesianxr7.github.io/bellwether"
Issues = "https://github.com/cartesianxr7/bellwether/issues"

[project.scripts]
bellwether = "bellwether.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
bellwether = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "RUF"]
ignore = []

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
